Move Children to Container. Initial pass.
