+++ osu.Framework/Graphics/Drawable.cs
@@ -14,6 +14,7 @@ using osu.Framework.Lists;
 using osu.Framework.Timing;
 using OpenTK;
 using OpenTK.Graphics;
+using osu.Framework.Graphics.Containers;
 
 namespace osu.Framework.Graphics
 {
@@ -27,27 +28,7 @@ namespace osu.Framework.Graphics
         /// A name used to identify this Drawable internally.
         /// </summary>
         public virtual string Name => string.Empty;
-
-        private LifetimeList<Drawable> children;
-        private IEnumerable<Drawable> pendingChildren;
-
-        public virtual IEnumerable<Drawable> Children
-        {
-            get { return children; }
-            set
-            {
-                if (!IsLoaded)
-                    pendingChildren = value;
-                else
-                {
-                    Clear();
-                    Add(value);
-                }
-            }
-        }
-
-        internal List<Drawable> CurrentChildren => children.Current;
-
+        
         private LifetimeList<ITransform> transforms;
 
         /// <summary>
@@ -462,14 +443,9 @@ namespace osu.Framework.Graphics
             }
         }
 
-        protected Drawable Parent { get; private set; }
+        public Drawable Parent { get; set; }
 
         protected virtual IComparer<Drawable> DepthComparer => new DepthComparer();
-
-        protected Drawable()
-        {
-            children = new LifetimeList<Drawable>(DepthComparer);
-        }
 
         /// <summary>
         /// Checks if this drawable is a child of parent regardless of nesting depth.
@@ -515,74 +491,7 @@ namespace osu.Framework.Graphics
             return false;
         }
 
-        public virtual Drawable Add(Drawable drawable)
-        {
-            if (drawable == null)
-                return null;
-
-            drawable.changeParent(this);
-            children.Add(drawable);
-
-            return drawable;
-        }
-
-        public void Add(IEnumerable<Drawable> collection)
-        {
-            foreach (Drawable d in collection)
-                Add(d);
-        }
-
-        public virtual bool Remove(Drawable p, bool dispose = true)
-        {
-            if (p == null)
-                return false;
-
-            bool result = children.Remove(p);
-            p.Parent = null;
-
-            if (dispose && p.IsDisposable)
-                p.Dispose();
-            else
-                p.Invalidate();
-
-            return result;
-        }
-
-        public int RemoveAll(Predicate<Drawable> match, bool dispose = true)
-        {
-            List<Drawable> toRemove = children.FindAll(match);
-            for (int i = 0; i < toRemove.Count; i++)
-                Remove(toRemove[i]);
-
-            return toRemove.Count;
-        }
-
-        public void Remove(IEnumerable<Drawable> range, bool dispose = true)
-        {
-            if (range == null)
-                return;
-
-            foreach (Drawable p in range)
-            {
-                if (p.IsDisposable)
-                    p.Dispose();
-                Remove(p);
-            }
-        }
-
-        public virtual void Clear(bool dispose = true)
-        {
-            foreach (Drawable t in children)
-            {
-                if (dispose)
-                    t.Dispose();
-                t.Parent = null;
-            }
-
-            children.Clear();
-
-            Invalidate(Invalidation.ScreenSpaceQuad);
-        }
+        
 
         protected virtual Quad DrawQuadForBounds => DrawQuad;
 
@@ -655,7 +564,7 @@ namespace osu.Framework.Graphics
 
         private List<DrawNode> validDrawNodes = new List<DrawNode>();
 
-        internal DrawNode GenerateDrawNodeSubtree(DrawNode node = null)
+        internal virtual DrawNode GenerateDrawNodeSubtree(DrawNode node = null)
         {
             if (node == null)
             {
@@ -671,50 +580,6 @@ namespace osu.Framework.Graphics
                 validDrawNodes.Add(node);
             }
 
-            if (children.Current.Count > 0)
-            {
-                if (node.Children != null)
-                {
-                    var current = children.Current;
-                    var target = node.Children;
-
-                    int j = 0;
-                    for (int i = 0; i < current.Count; i++)
-                    {
-                        if (!current[i].IsVisible) continue;
-
-                        if (j < target.Count && target[j].Drawable == current[i])
-                        {
-                            current[i].GenerateDrawNodeSubtree(target[j]);
-                        }
-                        else
-                        {
-                            if (j < target.Count)
-                                target.RemoveAt(j);
-                            target.Insert(j, current[i].GenerateDrawNodeSubtree());
-                        }
-
-                        j++;
-                    }
-
-                    if (j < target.Count)
-                        target.RemoveRange(j, target.Count - j);
-
-                }
-                else
-                {
-                    node.Children = new List<DrawNode>(children.Current.Count);
-
-                    foreach (Drawable child in children.Current)
-                        if (child.IsVisible)
-                            node.Children.Add(child.GenerateDrawNodeSubtree());
-                }
-            }
-            else
-            {
-                node.Children?.Clear();
-            }
-
             return node;
         }
 
@@ -726,21 +591,6 @@ namespace osu.Framework.Graphics
 
         protected virtual DrawNode CreateDrawNode() => new DrawNode();
 
-        /// <summary>
-        /// Perform any layout changes just before autosize is calculated.		
-        /// </summary>		
-        protected virtual void UpdateLayout()
-        {
-        }
-
-        /// <summary>
-        /// Updates the life status of children according to their IsAlive property.
-        /// </summary>
-        /// <returns>True iff the life status of at least one child changed.</returns>
-        protected virtual bool UpdateChildrenLife()
-        {
-            return children.Update();
-        }
 
         internal virtual void UpdateSubTree()
         {
@@ -754,13 +604,6 @@ namespace osu.Framework.Graphics
 
             Update();
             OnUpdate?.Invoke();
-
-            UpdateChildrenLife();
-
-            foreach (Drawable child in children.Current)
-                child.UpdateSubTree();
-
-            UpdateLayout();
         }
 
         protected virtual void Update()
@@ -777,26 +620,23 @@ namespace osu.Framework.Graphics
             return false;
         }
 
-        private void changeParent(Drawable parent)
+        internal void ChangeParent(Drawable parent)
         {
             if (Parent == parent)
                 return;
 
-            Parent?.Remove(this, false);
+            (Parent as Container)?.Remove(this, false);
             Parent = parent;
 
-            changeRoot(Parent?.Game);
+            ChangeRoot(Parent?.Game);
         }
 
-        private void changeRoot(Game root)
+        internal virtual void ChangeRoot(Game root)
         {
             if (root == null) return;
 
             Game = root;
             clockBacking.Invalidate();
-
-            foreach (Drawable c in children)
-                c.changeRoot(root);
         }
 
         /// <summary>
@@ -840,12 +680,6 @@ namespace osu.Framework.Graphics
 
         public virtual void Load()
         {
-            if (pendingChildren != null)
-            {
-                Add(pendingChildren);
-                pendingChildren = null;
-            }
-
             loaded = true;
             Invalidate();
         }
@@ -914,20 +748,6 @@ namespace osu.Framework.Graphics
             if (alreadyInvalidated || !shallPropagate)
                 return !alreadyInvalidated;
 
-            if (children != null)
-            {
-                foreach (var c in children)
-                {
-                    Debug.Assert(c != source);
-
-                    Invalidation childInvalidation = invalidation;
-                    //if (c.SizeMode == InheritMode.None)
-                    childInvalidation = childInvalidation & ~Invalidation.SizeInParentSpace;
-
-                    c.Invalidate(childInvalidation, this);
-                }
-            }
-
             return !alreadyInvalidated;
         }
 
@@ -992,9 +812,6 @@ namespace osu.Framework.Graphics
         public virtual Drawable Clone()
         {
             Drawable thisNew = (Drawable)MemberwiseClone();
-
-            thisNew.children = new LifetimeList<Drawable>(DepthComparer);
-            children.ForEach(c => thisNew.children.Add(c.Clone()));
 
             if (transforms != null)
             {
