+++ osu.Framework/Graphics/Drawable_TransformationHelpers.cs
@@ -20,20 +20,17 @@ namespace osu.Framework.Graphics
             transforms?.Clear();
         }
 
-        public Drawable Delay(double duration, bool propagateChildren = false)
+        public virtual Drawable Delay(double duration, bool propagateChildren = false)
         {
             if (duration == 0) return this;
 
             transformationDelay += duration;
-            if (propagateChildren)
-                Children.ForEach(c => c.Delay(duration, propagateChildren));
             return this;
         }
 
-        public Drawable DelayReset()
+        public virtual Drawable DelayReset()
         {
             Delay(-transformationDelay);
-            Children.ForEach(c => c.DelayReset());
             return this;
         }
